requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cssselect==1.2.0
darkdetect==0.5.1
fake-headers==1.0.2
fake-useragent==1.1.1
gevent==22.10.2
greenlet==2.0.1
grequests==0.6.0
html5lib==1.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
jellyfish==0.9.0
keyboard==0.13.5
lxml==4.9.2
MailTm==0.0.81
mouse==0.7.1
orjson==3.8.6
packaging==22.0
parse==1.19.0
Pillow==9.3.0
poe-api==0.2.4
pycparser==2.21
pyee==8.2.2
pyperclip==1.8.2
pyppeteer==1.0.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyquery==1.4.3
PySocks==1.7.1
pytesseract==0.3.10
python-socks==2.2.0
pywin32==305
PyYAML==6.0
requests==2.28.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tendo==0.3.0
tqdm==4.64.1
urllib3==1.26.13
w3lib==2.1.1
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
